# Optimize FFI Rust side type checks